exynos: Switch to a new versioned path and BSP variant setting
and add soong namespacing.

Copy-paste explanation:

There are two sets of BSP sources - These new ones intended for Android
13, and older ones originally intended for Android 10. The older ones
have support for some legacy things like HWC1 for older devices, and
the newer ones have support for newer things like gralloc4.

To keep support for older devices going, hoping that would continue to
be possibke, I want to keep the older sources alongside the newer ones.
But they can't co-exist as-is. Some changes are needed:
 - a change of paths to the BSP sources, which are referenced directly
   within them
 - a new BSP variant setting ('TARGET_SLSI_VARIANT' flag)
 - and finally, some soong namespacing.

The BSP variant setting and paths are the biggest challenge. For the
existing BSP sources, there's absolutely no versioning. It's the same
old generic "linaro"/"samsung_slsi-linaro" for both sets of BSP sources.
Sadly, my expressions of concern for my life to not be made more hellish
were quite badly dismissed, all the while I'd think it would be best
practice to version things like this to not keep using the same old
generic identifiers and not run into issues (or for others to not do so)
later, so I'm on my own to figure out what to do here.

My solution that touches the least I can is as follows:

For Android 13 - The older Android 10 BSP sources are typically used,
and since I'd like to use 13 sources as I have a device that requires
gralloc4, I need to make changes within them.
To manage it, the naming is as follows:
  Directory for 10 sources - 'samsung_slsi-linaro'
  Directory for 13 sources - 'samsung_slsi-linaro_13-e850-96'
  Git branch name for 10 sources - '...'
  Git branch name for 13 sources - '..._13-e850-96'
  SLSI variant flag setting for 10 sources - 'linaro'
  SLSI variant flag setting for 13 sources - 'linaro_13-e850-96'
The only changes here are for 13 sources. Naming is left untouched for
10 sources so as to not break anything for trees using them currently.

For Android 14+ - I'd want to preserve support for older devices,
provided if, and hoping that, running newer Android versions doesn't
become an impossibility for them, so, while the newer Android 13 BSP
sources are typically used, I'd have to make changes to the 10 sources
to keep them alongside.
To manage it, the naming will be as follows:
  Directory for 10 sources - 'samsung_slsi-linaro_10-e850-96'
  Directory for 13 sources - 'samsung_slsi-linaro'
  Git branch name for 10 sources - '..._10-e850-96'
  Git branch name for 13 sources -  '...' / '..._13-e850-96'
  SLSI variant flag setting for 10 sources - 'linaro_10-e850-96'
  SLSI variant flag setting for 13 sources - 'linaro'
The directory and variant flag setting will be untouched for 13 sources,
so as to stay in line with everyone else, and changes would only need to
be made within 10 sources. However, for consistency, there will always
be versioned git branches in future, alongside non-versioned ones with
whichever sources are typically used, though this all remains less than
ideal in any case.

Lastly, soong namespacing is added specifically for the "exynos" and
"graphics" parts of the sources. Without, there are conflicts between
the two sets of sources. For all devices, BoardConfigCommon.mk in the
collection of BSP configs is included, and namespaces can be added
there, and also the SLSI variant flag setting can be changed, for a
clean result where no changes are required in device trees at all.

Change-Id: I49efeac8ee68b052ef48faeedeab1628bf1f78ef